Fixed double frees
The style assignment was freeing the style tag twice, and you can't move
elements from one document to the next.

libxml2 fucks up references, only letting you set or clear the parent,
child and sibling pointers, not the (to and from) document pointer or
the pointer to a namespace (which points to a document). When the
document is freed, all things that point to it are freed, thus
xmlFreeNode will always cause a double free even when you use
xmlUnlinkNode beforehand. At least that's what I gather from eyeballing
the libxml2 code.

For that matter, moving a node from one document to a next will add
references to it in the new document, leaving the ones in the old
document, so when the two documents are freed... another unavoidable
double free. There's a function to "adopt" a node moving it with
references adjusted properly, but it's labeled as experimental. So just
copying the node is easier.

Of course when you copy a node, the copy also is referenced by the old
document, so adding it to the new document still causes a double free.
Only if you use xmlDocCopyNode will it copy the old node, but assign a
new document to the copy.

Why is libxml2 so ugh...