style and stylesheet?
Allow setting a stylesheet inline in the text file via the <stylesheet>
tag. Should I use a <link rel="stylesheet"> tag instead? Can't use a
<style> tag because that is a placeholder for where to put the link to
the document's style sheet. Though since that link always goes in the
head and its position in head doesn't matter, there's no need for a
placeholder for it in the template. Maybe in the future <style> will
parse out inline styles, put them in the style sheet and make sure
there's a link to it.