Added a DOCTYPE clause oops firefox breaks
Added code for making sure a DOCTYPE is in there so that entities like
&nbsp; don't produce invalid XML. Then commented out all that work
because Firefox is broken. If it sees a DOCTYPE clause, it ignores the <?xml?> processing instruction and discards any encoding information, forcing the document to view as latin-1 decoded. This is wrong. If anyone knows how to submit a bug to Firefox go ahead, but I gave up on them long ago. Have to serve them, no voice in changing them. Thus all that is left to do, is TRICKERY. Using TRICKERY one can omit the DOCTYPE declaration and just pretend that it's valid XML, or use the utf-8 equivalent to any XML entity and remove all entities besides the builtin ones (amp, gt, lt).