Firefox doesn't suck yay also load entities
I was wrong, Firefox doesn't suck my web server does. I thought it was
ignoring <?xml?> because of the DOCTYPE declaration, but it really was
because I accidentally used the .html extension instead of .xhtml. With
the proper MIME type, Firefox respects the XML processing instruction
for character encoding just fine.

Also made a thing to load DTDs and cache them locally. Funnily enough,
the HTTP client included in libxml2 makes the location of almost all
DTDs crash (w3.org). I assume it's because w3.org's stupid CGI script
that they should never be using because loading files statically is a
hundred times more efficient, assumes HTTP/1.1 and dies when it's not.

Easy enough to hack a call to the curl program in there. Which is itself
easier (and really more secure) than using libcurl.

It's easy enough to load cache files via a libxml2 input handler, but
getting the network resource into the cache file itself... bleh. Anyway
it works.